Moved asset components to own folder
