+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/assets-picker/assets-list-view.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 var StaticAssetItemView = require('./static-asset-item-view');
-var StaticAssetsCollection = require('../../../../../data/static-assets-collection');
+var StaticAssetsCollection = require('../../../../../../data/static-assets-collection');
 var template = require('./assets-list-view.tpl');
 
 module.exports = CoreView.extend({
