+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/assets-picker/assets-list-view.tpl
