+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/assets-picker/assets-view.js
@@ -1,12 +1,12 @@
 var CoreView = require('backbone/core-view');
 var template = require('./assets-view.tpl');
-var ScrollView = require('../../../../scroll/scroll-view');
+var ScrollView = require('../../../../../scroll/scroll-view');
 
 var AssetsListView = require('./assets-list-view');
 
-var MakiIcons = require('./assets/maki-icons');
-var PinIcons = require('./assets/pin-icons');
-var SimpleIcons = require('./assets/simple-icons');
+var MakiIcons = require('../assets/maki-icons');
+var PinIcons = require('../assets/pin-icons');
+var SimpleIcons = require('../assets/simple-icons');
 
 module.exports = CoreView.extend({
   className: 'Dialog-content Dialog-content--expanded',
